feat: add admin view for pending transfer requests with shared table component
